adicionado configuração do NA e NF (applicacao)

+++ components/utils/include/project_config.h
@@ -154,8 +154,8 @@ typedef struct __attribute__((__packed__))
 {
     char contactor[50];       /*!< Contactor type */ //todo: usar isso nas classes
     char pressure[50];   /*!< Pressure switch type *///todo usar isso nas classes
-    uint16_t pressurization_time;   /*!< Pressurization time */
-    uint8_t on_off_time;            /*!< On/off time */
+    uint32_t pressurization_time;   /*!< Pressurization time */
+    uint16_t on_off_time;            /*!< On/off time */
 } pivot_config;
 
 /**
